chore(*): remove unused files
