--- frontend/src/pages/np/components/LatestFeed.tsx
@@ -1,261 +1,0 @@
-import { NanopubIcon } from "@/components/nanopub-icon";
-import { PaginationControls } from "@/components/pagination-controls";
-import { RelativeDateTime } from "@/components/relative-datetime";
-import { useTheme } from "@/components/theme-provider";
-import { AsyncLabel } from "@/hooks/use-labels";
-import { usePagination } from "@/hooks/use-pagination";
-import { toScienceLiveNPUri } from "@/lib/uri";
-import {
-  FEED_GROUPS,
-  FEED_TEMPLATE_LABELS,
-  type FeedResult,
-  type FeedTemplateKey,
-  useFeed,
-} from "@/pages/feed/use-feed";
-import { Check, FilterX, Loader2, Minus } from "lucide-react";
-import { useCallback, useMemo, useState } from "react";
-import { Link } from "react-router-dom";
-import {
-  TEMPLATE_METADATA,
-  TEMPLATE_URI,
-  getTemplateColorClass,
-} from "../create/components/templates/registry-metadata";
-import { TEMPLATE_VIEW_ICONS } from "../view/view-registry";
-import { NanopubTemplateIcon } from "./SearchResultList";
-
-/** Initial checked state for each template key — false means unchecked */
-const INITIAL_CHECKED: Record<FeedTemplateKey, boolean> = {
-  AIDA_SENTENCE: true,
-  CITATION_CITO: true,
-  ANNOTATE_QUOTATION: false,
-  COMMENT_PAPER: false,
-  GEO_COVERAGE: false,
-  DATASET: false,
-  RESEARCH_SOFTWARE: false,
-  ODRL_POLICY: false,
-  ODRL_ACCESS_GRANT: false,
-  PICO_RESEARCH_QUESTION: true,
-  PCC_RESEARCH_QUESTION: true,
-  PRISMA_SEARCH_STRATEGY: true,
-  PRISMA_DATABASE_SEARCH: false,
-  PRISMA_SEARCH_EXECUTION_DATASET: false,
-  PRISMA_STUDY_INCLUSION: false,
-  PRISMA_STUDY_ASSESSMENT: false,
-  PRISMA_FULL_SCREENING: false,
-  FORRT_CLAIM: false,
-  FORRT_REPLICATION: false,
-  FORRT_REPLICATION_OUTCOME: true,
-  FORRT_KL_REPLICATION: false,
-  FORRT_KL_REPLICATION_OUTCOME: true,
-  RESEARCH_SYNTHESIS: false,
-};
-
-export function LatestFeed() {
-  const { resolvedTheme } = useTheme();
-
-  const [checked, setChecked] =
-    useState<Record<FeedTemplateKey, boolean>>(INITIAL_CHECKED);
-
-  const { currentPage, setPage, resetPage } = usePagination();
-
-  const selected = useMemo(() => {
-    const s = new Set<FeedTemplateKey>();
-    for (const [key, val] of Object.entries(checked)) {
-      if (val) s.add(key as FeedTemplateKey);
-    }
-    return s;
-  }, [checked]);
-
-  const { results, loading, error, hasMore } = useFeed({
-    selectedKeys: selected,
-    page: currentPage,
-  });
-
-  const toggle = useCallback(
-    (key: FeedTemplateKey) => {
-      setChecked((prev) => ({ ...prev, [key]: !prev[key] }));
-      resetPage();
-    },
-    [resetPage],
-  );
-
-  const clearFilters = useCallback(() => {
-    setChecked((prev) => {
-      const next = { ...prev };
-      for (const key of Object.keys(next) as FeedTemplateKey[]) {
-        next[key] = false;
-      }
-      return next;
-    });
-    resetPage();
-  }, [resetPage]);
-
-  const toggleGroup = useCallback(
-    (keys: FeedTemplateKey[]) => {
-      setChecked((prev) => {
-        const allOn = keys.every((k) => prev[k]);
-        const next = { ...prev };
-        for (const k of keys) {
-          next[k] = !allOn;
-        }
-        return next;
-      });
-      resetPage();
-    },
-    [resetPage],
-  );
-
-  return (
-    <div className="flex flex-col gap-6 lg:flex-row">
-      {/* Sidebar: template filters */}
-      <aside className="flex w-full flex-col gap-4 lg:w-64 lg:min-w-64">
-        <div className="flex items-center justify-between">
-          <h2 className="text-sm font-medium text-muted-foreground">
-            Filter by template
-          </h2>
-          {selected.size > 0 && (
-            <button
-              type="button"
-              onClick={clearFilters}
-              className="flex items-center gap-1 text-xs text-muted-foreground hover:text-foreground transition-colors"
-            >
-              <FilterX className="h-3.5 w-3.5" />
-              Clear filters
-            </button>
-          )}
-        </div>
-        {FEED_GROUPS.map((group) => {
-          const allOn = group.keys.every((k) => checked[k]);
-          const someOn = !allOn && group.keys.some((k) => checked[k]);
-          return (
-            <div key={group.label} className="flex flex-col gap-1.5">
-              <label
-                className="flex cursor-pointer items-center gap-2 text-left select-none"
-                onClick={() => toggleGroup(group.keys)}
-              >
-                <FeedCheckbox
-                  state={
-                    allOn ? "checked" : someOn ? "indeterminate" : "unchecked"
-                  }
-                />
-                <span className="text-sm font-medium">{group.label}</span>
-              </label>
-              <div className="ml-6 flex flex-col gap-1">
-                {group.keys.map((key) => {
-                  const uri = TEMPLATE_URI[key];
-                  const Icon = TEMPLATE_VIEW_ICONS[uri];
-                  const color = TEMPLATE_METADATA[uri]?.color;
-                  return (
-                    <label
-                      key={key}
-                      className="flex cursor-pointer items-center gap-2 text-sm font-normal select-none"
-                      onClick={() => toggle(key)}
-                    >
-                      <FeedCheckbox
-                        state={checked[key] ? "checked" : "unchecked"}
-                      />
-                      {Icon ? (
-                        <Icon
-                          className={`w-3.5 h-3.5 min-w-3.5 min-h-3.5 ${getTemplateColorClass(color, resolvedTheme)}`}
-                        />
-                      ) : (
-                        <NanopubIcon className="w-2.5 h-2.5 min-w-2.5 min-h-2.5 text-muted-foreground" />
-                      )}
-                      {FEED_TEMPLATE_LABELS[key]}
-                    </label>
-                  );
-                })}
-              </div>
-            </div>
-          );
-        })}
-      </aside>
-
-      {/* Results */}
-      <section className="flex-1">
-        {loading && (
-          <div className="flex items-center gap-2 text-sm text-muted-foreground">
-            <Loader2 className="h-4 w-4 animate-spin" />
-            Loading...
-          </div>
-        )}
-        {error && <p className="text-sm text-destructive">Error: {error}</p>}
-        {!loading && !error && results.length === 0 && (
-          <p className="text-sm text-muted-foreground">
-            {selected.size === 0
-              ? "No nanopublications found."
-              : "No nanopublications found for the selected templates."}
-          </p>
-        )}
-        {!loading && results.length > 0 && (
-          <>
-            <FeedResultList results={results} />
-            <PaginationControls
-              currentPage={currentPage}
-              hasMore={hasMore}
-              loading={loading}
-              onPageChange={setPage}
-            />
-          </>
-        )}
-      </section>
-    </div>
-  );
-}
-
-function FeedCheckbox({
-  state,
-}: {
-  state: "checked" | "unchecked" | "indeterminate";
-}) {
-  return (
-    <span
-      className={`flex size-4 shrink-0 items-center justify-center rounded-lg border shadow-xs ${
-        state === "unchecked"
-          ? "border-input dark:bg-input/30"
-          : "border-primary bg-primary text-primary-foreground"
-      }`}
-    >
-      {state === "checked" && <Check className="size-3.5" />}
-      {state === "indeterminate" && <Minus className="size-3.5" />}
-    </span>
-  );
-}
-
-function FeedResultList({ results }: { results: FeedResult[] }) {
-  return (
-    <div className="flex flex-col gap-3">
-      {results.map((r, i) => {
-        const displayLabel = r.description || r.label;
-        return (
-          <div
-            key={r.np || i}
-            className="flex flex-col gap-2 rounded-lg border bg-card p-4"
-          >
-            <Link to={toScienceLiveNPUri(r.np)}>
-              <div className={`font-medium flex flex-row`}>
-                <NanopubTemplateIcon template={r.template} />
-                {displayLabel || "Untitled Nanopublication"}
-              </div>
-            </Link>
-            {/* {r.description && r.label && (
-              <p className="text-xs text-muted-foreground line-clamp-1">
-                {r.label}
-              </p>
-            )} */}
-            <div className="flex items-center gap-2 text-sm text-muted-foreground">
-              <span className="text-xs truncate">
-                By <AsyncLabel uri={r.creator} link />
-              </span>
-            </div>
-            {r.date && (
-              <div className="flex items-center gap-2 text-xs text-muted-foreground">
-                <RelativeDateTime date={new Date(r.date)} />
-              </div>
-            )}
-          </div>
-        );
-      })}
-    </div>
-  );
-}
